Fix time bug

+++ include/core/bot.hpp
@@ -42,7 +42,7 @@ class ChessBot
         Move move;
         int move_idx;
         bool is_root_search;
-        int mate_in;
+        int64_t mate_in;
     };
 
     ChessBoardEvaluator evaluator;
@@ -69,7 +69,7 @@ class ChessBot
         return board->GetZobristHash();
     }
 
-    Evaluation MainSearch(Evaluation alpha, Evaluation beta, int depth, int ply, int& mate_in);
+    Evaluation MainSearch(Evaluation alpha, Evaluation beta, int depth, int ply, int64_t& mate_in);
     Evaluation SearchCore(SearchParams& params);
 
     int DepthExtension(const Move& move);
